requirements/requirements_local.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements_local.txt base.in local.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
astroid==3.3.11
    # via pylint
black==25.12.0
    # via -r local.in
bookmark-it @ file:///C:/Users/COTTRILLAD/Documents/1work/Python/djcode/apps/bookmark_it/dist/bookmark-it-2024.3.8.tar.gz#sha256=413ca50567ec6d9a353a9e93d9c0755ad77134653ae7a5e7704b897779d41bb9
    # via -r base.in
build==1.3.0
    # via pip-tools
click==8.2.1
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   pylint
    #   pytest
coverage[toml]==7.13.0
    # via
    #   -r local.in
    #   pytest-cov
dill==0.4.0
    # via pylint
django==5.2.9
    # via
    #   -r base.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-leaflet
    #   django-taggit
    #   djangorestframework
    #   drf-yasg
django-colorfield==0.14.0
    # via -r base.in
django-cors-headers==4.7.0
    # via -r base.in
django-debug-toolbar==6.1.0
    # via -r local.in
django-extensions==4.1
    # via -r local.in
django-filter==25.2
    # via -r base.in
django-leaflet==0.33.0
    # via -r base.in
django-taggit==6.1.0
    # via -r base.in
djangorestframework==3.16.1
    # via
    #   -r base.in
    #   drf-excel
    #   drf-yasg
drf-excel==2.5.3
    # via drf-renderer-xlsx
drf-renderer-xlsx==1.0.0
    # via -r base.in
drf-yasg==1.21.11
    # via -r base.in
et-xmlfile==2.0.0
    # via openpyxl
execnet==2.1.1
    # via pytest-xdist
factory-boy==3.3.0
    # via -r local.in
faker==37.6.0
    # via factory-boy
gdal @ file:///C:/Users/COTTRILLAD/Downloads/GDAL-3.9.2-cp312-cp312-win_amd64.whl
    # via -r base.in
inflection==0.5.1
    # via drf-yasg
iniconfig==2.1.0
    # via pytest
isort==6.1.0
    # via
    #   -r local.in
    #   pylint
markdown==3.10
    # via -r base.in
markdown2==2.5.4
    # via -r base.in
mccabe==0.7.0
    # via pylint
mypy-extensions==1.1.0
    # via black
openpyxl==3.1.5
    # via
    #   -r base.in
    #   drf-excel
packaging==25.0
    # via
    #   black
    #   build
    #   drf-yasg
    #   pytest
pathspec==0.12.1
    # via black
pillow==12.0.0
    # via
    #   -r base.in
    #   django-colorfield
pip-tools==7.5.0
    # via -r local.in
platformdirs==4.4.0
    # via
    #   black
    #   pylint
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg2-binary==2.9.10
    # via -r base.in
pygments==2.19.2
    # via pytest
pylint==3.3.9
    # via
    #   -r local.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.6.1
    # via -r local.in
pylint-plugin-utils==0.9.0
    # via
    #   -r local.in
    #   pylint-django
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.2
    # via
    #   -r local.in
    #   pytest-cov
    #   pytest-django
    #   pytest-xdist
pytest-cov==7.0.0
    # via -r local.in
pytest-django==4.11.1
    # via -r local.in
pytest-xdist==3.8.0
    # via -r local.in
pytokens==0.3.0
    # via black
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.2
    # via drf-yasg
resource-library @ file:///c:/Users/COTTRILLAD/Documents/1work/Python/djcode/apps/resource_library/dist/resource-library-2022.6.29.tar.gz#sha256=9ae13190bf377fb49438767053aba2227d4058b1e5f21a631106bf9262e6dfd2
    # via -r base.in
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
ticket-tracker @ file:///C:/Users/COTTRILLAD/Documents/1work/Python/djcode/apps/tickettracker2/dist/ticket-tracker-2024.4.3.tar.gz#sha256=c0c804f3cba97171a3804412ed66d1c36c97cf37537a512c83098c51fae8c880
    # via -r base.in
tomlkit==0.13.3
    # via pylint
tzdata==2025.2
    # via
    #   django
    #   faker
uritemplate==4.2.0
    # via drf-yasg
wheel==0.45.1
    # via pip-tools
whitenoise==6.11.0
    # via -r local.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
